Move options files to their own directory
Also remove thhe now unused enums file
