--- DfE.FindInformationAcademiesTrusts/CookieStatusEnums.cs
@@ -1,13 +1,0 @@
-namespace DfE.FindInformationAcademiesTrusts;
-
-public class CookieStatusEnums
-{
-    public enum CookieStatus
-    {
-        NotSelected = 0,
-
-        Accepted = 1,
-
-        Rejected = 2
-    }
-}
