+++ DfE.FindInformationAcademiesTrusts/Options/TestOverrideOptions.cs
@@ -1,0 +1,8 @@
+namespace DfE.FindInformationAcademiesTrusts.Options;
+
+public class TestOverrideOptions
+{
+    public const string ConfigurationSection = "TestOverride";
+
+    public string? PlaywrightTestSecret { get; init; }
+}
