+++ DfE.FindInformationAcademiesTrusts/Pages/_ViewImports.cshtml
@@ -1,4 +1,5 @@
 ﻿@using DfE.FindInformationAcademiesTrusts
+@using DfE.FindInformationAcademiesTrusts.Options
 @using Microsoft.ApplicationInsights.AspNetCore
 @using Microsoft.Extensions.Options
 @namespace DfE.FindInformationAcademiesTrusts.Pages
