--- DfE.FindInformationAcademiesTrusts/TestOverrideOptions.cs
@@ -1,8 +1,0 @@
-namespace DfE.FindInformationAcademiesTrusts;
-
-public class TestOverrideOptions
-{
-    public const string ConfigurationSection = "TestOverride";
-
-    public string? PlaywrightTestSecret { get; init;}
-}
